back/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.5
Click==7.0
cryptography==3.4.7
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.11.1
pycparser==2.20
PyJWT==2.0.1
pyrsistent==0.17.3
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.13
Werkzeug==0.16.0
WTForms==2.2.1
